Move task queue to separate class and add Timeout and Interval.
